Fix offByOne-error
This fix works, but I didn't find the root cause. I can find nothing wrong with
my array to buffer transformation, but somewhere blocksPerLine is not correctly
used. I guess the transformation revealed this error. For now, this is better
than nothing, but the cause should be found and eliminated